+++ src/main/java/it/gov/pagopa/payhub/activities/connector/classification/ClassificationService.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface ClassificationService {
     Integer saveAll(List<Classification> classificationDTOList);
-    Long deleteByOrganizationIdAndIufAndLabel(Long organizationId, String iuf, ClassificationsEnum classification);
-    Long deleteBySemanticKey(TransferSemanticKeyDTO transferSemanticKeyDTO);
-    Long deleteByOrganizationIdAndIudAndLabel(Long organizationId, String iud, ClassificationsEnum classification);
+    Integer deleteByOrganizationIdAndIufAndLabel(Long organizationId, String iuf, ClassificationsEnum classification);
+    Integer deleteBySemanticKey(TransferSemanticKeyDTO transferSemanticKeyDTO);
+    Integer deleteByOrganizationIdAndIudAndLabel(Long organizationId, String iud, ClassificationsEnum classification);
 }
